+++ WordPressKitTests/AccountSettingsRemoteTests.swift
@@ -90,7 +90,7 @@ class AccountSettingsRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })
 
@@ -107,7 +107,7 @@ class AccountSettingsRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
             expect.fulfill()
         })
 
@@ -260,7 +260,7 @@ class AccountSettingsRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })
 
@@ -293,7 +293,7 @@ class AccountSettingsRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.invalidInput.rawValue, "The error code should be 0 - invalid input")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.invalidInput.rawValue, "The error code should be 0 - invalid input")
             expect.fulfill()
         })