+++ WordPressKitTests/PeopleServiceRemoteTests.swift
@@ -104,7 +104,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
             expect.fulfill()
         })
 
@@ -121,7 +121,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })
 
@@ -138,7 +138,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.invalidInput.rawValue, "The error code should be 0 - invalid input")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.invalidInput.rawValue, "The error code should be 0 - invalid input")
             expect.fulfill()
         })
 
@@ -170,7 +170,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })
 
@@ -204,7 +204,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })
 
@@ -221,7 +221,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
             expect.fulfill()
         })
 
@@ -253,7 +253,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })
 
@@ -327,7 +327,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
             // Then
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.authorizationRequired.rawValue, "The error code should be 2 - authorization required")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.authorizationRequired.rawValue, "The error code should be 2 - authorization required")
             expect.fulfill()
         })
 
@@ -347,7 +347,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
             // Then
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })
 
@@ -367,7 +367,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
             // Then
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })
         waitForExpectations(timeout: timeout, handler: nil)
@@ -386,7 +386,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
             // Then
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
             expect.fulfill()
         })
         waitForExpectations(timeout: timeout, handler: nil)
@@ -421,7 +421,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
             // Then
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })
 
@@ -486,7 +486,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
             expect.fulfill()
         })
 
@@ -503,7 +503,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })
 
@@ -553,7 +553,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })
 
@@ -570,7 +570,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.authorizationRequired.rawValue, "The error code should be 2 - authorization_required")
             expect.fulfill()
         })
 
@@ -648,7 +648,7 @@ class PeopleServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })