+++ WordPressKitTests/PlanServiceRemoteTests.swift
@@ -120,7 +120,7 @@ class PlanServiceRemoteTests: RemoteTestCase, RESTTestable {
         }, failure: { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         })
 
@@ -248,7 +248,7 @@ class PlanServiceRemoteTests: RemoteTestCase, RESTTestable {
         }) { error in
             let error = error as NSError
             XCTAssertEqual(error.domain, "WordPressKit.WordPressComRestApiError", "The error domain should be WordPressComRestApiError")
-            XCTAssertEqual(error.code, WordPressComRestApiError.unknown.rawValue, "The error code should be 7 - unknown")
+            XCTAssertEqual(error.code, WordPressComRestApiErrorCode.unknown.rawValue, "The error code should be 7 - unknown")
             expect.fulfill()
         }